chore(sdk-lib-mpc): rename files to *Proof
TICKET: HSM-102
