+++ modules/sdk-lib-mpc/src/tss/ecdsa/index.ts
@@ -1,5 +1,5 @@
 export * as EcdsaTypes from './types';
-export * as EcdsaRangeProof from './rangeproof';
+export * as EcdsaRangeProof from './rangeProof';
 export * as EcdsaPaillierProof from './paillierProof';
 export * as EcdsaPaillierBlumProof from './paillierBlumProof';
